investment-agent/requirements.txt
pydantic==2.10.1
phidata
ollama
yfinance
duckduckgo-search
chainlit